Distinguish between composite and simple tabulate.
The difference places restrictions on the composite to not allow
simultaneous centering and interpolation to different qvalues.
The restriction is in place because from the experimental feature
vector it is difficult to determine what section should be used
as the new qvalues. Therefore require user to interpolate before
and simply allow centering in composite.